multiple ways of handling json and converting to java object
